Fix PDP 404 pages
